separate geoseries trait
